Major refactor of table logic
Table directives don't work if the table a child of a list item, so I refactored
the table logic to output simple tables. It turns out that the subset of the
"advanced" directive-based functionality is entirely supported by the simple
tables, so no functionality was lost due to the conversion.